Move functions for default nature and ability to `starter-select-ui-utils.ts`, and used `getStarterData` so fresh start is applied
