+++ src/data/challenge.ts
@@ -900,8 +900,14 @@ export class FreshStartChallenge extends Challenge {
     // Remove cost reduction
     starterDataEntry.valueReduction = 0;
 
-    // Always start with neutral nature
-    dexEntry.natureAttr = Nature.HARDY;
+    // Remove natures except for the default ones
+    const neutralNaturesAttr =
+      (1 << (Nature.HARDY + 1))
+      | (1 << (Nature.DOCILE + 1))
+      | (1 << (Nature.SERIOUS + 1))
+      | (1 << (Nature.BASHFUL + 1))
+      | (1 << (Nature.QUIRKY + 1));
+    dexEntry.natureAttr &= neutralNaturesAttr;
 
     // Cap all ivs at 15
     for (let i = 0; i < 6; i++) {
